refactored code to fit-in
